requirements.txt
requests
readability-lxml
bs4